chess quant v2/chessquant_ml_research_sandbox_v2/chessquant_ml_research_sandbox_v2/pyproject.toml
[tool.poetry]
name = "chessquant-ml-research-sandbox-v2"
version = "0.1.0"
description = "Research sandbox for ChessQuant model experiments and policy evaluation"
authors = ["John Scott <chatgpt4.mm.acc.2@gmail.com>"]
packages = [{ include = "chessquant_ml", from = "src" }]

[tool.poetry.dependencies]
python = "^3.13"
numpy = "^2.1.0"
pandas = "^2.2.0"
pyarrow = "^18.0.0"
scikit-learn = "^1.5.0"
xgboost = "^2.1.0"
matplotlib = "^3.9.0"
pyyaml = "^6.0.2"
joblib = "^1.4.2"
mlflow = "^2.17.0"
tqdm = "^4.67.3"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
ruff = "^0.6.7"

[tool.ruff]
line-length = 100
target-version = "py313"

[build-system]
requires = ["poetry-core>=1.8.0"]
build-backend = "poetry.core.masonry.api"
